Updated home screen location
